requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh upgrade
agno==2.5.15
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
croniter==6.2.2
cryptography==46.0.7
curl-cffi==0.15.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
duckdb==1.5.1
durationpy==0.10
email-validator==2.3.0
exa-py==2.11.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.1
fastar==0.10.0
filelock==3.25.2
flatbuffers==25.12.19
frozendict==2.4.7
fsspec==2026.3.0
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.10.1
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multitasking==0.0.12
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.31.0
openinference-instrumentation==0.1.46
openinference-instrumentation-agno==0.1.29
openinference-semantic-conventions==0.1.28
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
pandas==3.0.2
parallel-web==0.4.2
peewee==4.0.4
pgvector==0.4.2
pillow==12.2.0
platformdirs==4.9.6
proto-plus==1.27.2
protobuf==6.33.6
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pyasn1-modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pygithub==2.9.0
pygments==2.20.0
pyjwt==2.12.1
pynacl==1.6.2
pyparsing==3.3.2
pypdf==6.10.0
pypika==0.51.1
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
pytz==2026.1.post1
pyyaml==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
slack-sdk==3.41.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
yfinance==1.2.1
zipp==3.23.0
